refector after merge master

+++ src/main/java/com/example/nanuer_server/domain/entity/MyPageEntity.java
@@ -21,7 +21,7 @@ public class MyPageEntity extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "my_page_id")
-    private Long myPageId;
+    private int myPageId;
 
     //이름 수정
     @OneToOne(mappedBy = "myPageEntity")
